requirements.txt
numpy==1.23.4
opencv_python==4.7.0.68
Pillow==9.4.0
streamlit==1.17.0
tensorflow==2.10.0
